Start creating the tracking bot class
Add name, email and pathToQueryUser variables, and declare them as private.
Call initTracking method inside the constructor.
Add getter method to email variable.
Create initTracking method, to be called whem a new instance is created, and add the follower and following lists to the user's mongodb collection, as well as the time it was done.
Start creating checkGithubProfile method, call user current followers and following list from github api, and get the ones saved in database, needs to complete the method after that.