Add new methods and advance in checkGithubProfile method
setIsYouFollowing method verifies the list of follows and unfollows, and marks the isYouFollowing property based on the followingResponse data.
getFollowersListChanges filters the followersResponse and savedFollowers lists to find the new followers, or unfollows made.
Inside the initTracking method, map through the followers array to return just a object containing the necessary data, and avoid sending useless information to the database.
Call setIsYouFollowing and getFollowersListChanges inside checkGithubProfile method.