Refactor all methods
Github API has a limit of results per page when querying a user followers or followings, so before every query to the users' followers or followings, first query the user, to get its number of followers or followings, then assign to a page const the number of pages that will be needed to query, then use a for loop to query all the results, and push them to an array.
Still going to refactor and add comments to where this is being done.